[onert] Introduce common CalculateActivationRange
This commit introduces CalculateActivationRange in core util.
